requirements.txt
agno
python-dotenv
google-generativeai
lancedb
sqlite3